+++ admin/auth/logout.php
