+++ admin/user/edit_user.php
@@ -1,32 +1,41 @@
 <?php
 session_start();
-require '../model/connect.php';
+require '../../model/connect.php';
 
 if (!isset($_SESSION['admin'])) {
     header("Location: login.php");
     exit;
 }
 
+$id = $_GET['id'] ?? 0;
+$stmt = $conn->prepare("SELECT * FROM user WHERE id = ?");
+$stmt->execute([$id]);
+$user = $stmt->fetch();
+
+if (!$user) {
+    die("Sản phẩm không tồn tại!");
+}
+
 if ($_SERVER['REQUEST_METHOD'] == 'POST') {
-    $title = $_POST['title'];
-    $price = $_POST['price'];
-    $discount = $_POST['discount'];
-    $thumbnail = $_POST['thumbnail'];
-    $description = $_POST['description'];
-    $color = $_POST['color'];
-    $size = $_POST['size'];
+    $fullname = $_POST['fullname'];
+    $email = $_POST['email'];
+    $phone_number = $_POST['phone_number'];
+    $address = $_POST['address'];
 
-    if (!empty($title) && $price > 0) {
+    if (!empty($fullname) && $email > 0) {
+        // if ($image['size'] > 0) {
         // $imageName = time() . '_' . $image['name'];
-        // move_uploaded_file($image['tmp_name'], "assets/images/" . $thumbnail);
-
-        $stmt = $conn->prepare("INSERT INTO product (title, price,discount, thumbnail, description, color, size) VALUES (?, ?, ?, ?, ?, ?, ?)");
-        $stmt->execute([$title, $price, $discount, $thumbnail, $description, $color, $size]);
-
-        header("Location: product.php");
+        // move_uploaded_file($image['tmp_name'], "assets/images/" . $imageName);
+        $stmt = $conn->prepare("UPDATE user SET fullname=?, email=?,phone_number=?, address=? WHERE id=?");
+        $stmt->execute([$fullname, $email, $phone_number, $address, $id]);
+        // } else {
+        // $stmt = $conn->prepare("UPDATE products SET title=?, price=? WHERE id=?");
+        // $stmt->execute([$title, $price, $id]);
+        // }
+        header("Location: user_management.php");
         exit;
     } else {
-        $error = "Vui lòng nhập tin và tải lên hình ảnh!";
+        $error = "Vui lòng nhập đầy đủ tin!";
     }
 }
 ?>
@@ -35,7 +44,7 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
 <html lang="vi">
 
 <head>
-    <title>Thêm Sản phẩm</title>
+    <title>sửa Sản phẩm</title>
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css">
     <style>
         body {
@@ -181,63 +190,28 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
     </div>
 
     <div id="content">
-        <h2>Thêm Sản phẩm</h2>
+        <h2>Chỉnh sửa thông tin khách hàng</h2>
         <form method="POST" enctype="multipart/form-data" class="mt-3">
             <?php if (!empty($error)) echo "<p class='text-danger'>$error</p>"; ?>
             <div class="mb-3">
-                <label>Tên sản phẩm:</label>
-                <input type="text" name="title" class="form-control" required>
+                <label>Họ và tên:</label>
+                <input type="text" name="fullname" class="form-control" value="<?= $user['fullname'] ?>" required>
             </div>
             <div class="mb-3">
-                <label>Giá:</label>
-                <input type="number" name="price" class="form-control" required>
+                <label>email:</label>
+                <input type="email" name="email" class="form-control" value="<?= $user['email'] ?>" required>
             </div>
             <div class="mb-3">
-                <label>Giá discount:</label>
-                <input type="number" name="discount" class="form-control" required>
+                <label>Số điện thoại:</label>
+                <input type="number" name="phone_number" class="form-control" value="<?= $user['phone_number'] ?>" required>
             </div>
             <div class="mb-3">
-                <label>Hình ảnh:</label>
-                <input type="text" name="thumbnail" class="form-control" required>
+                <label>Địa chỉ:</label>
+                <input type="text" name="address" class="form-control" value="<?= $user['address'] ?>" required>
             </div>
-            <div class="mb-3">
-                <label>Mô tả:</label>
-                <input type="text" name="description" class="form-control" required>
-            </div>
-            <div class="mb-3">
-                <label>Màu sắc:</label>
-                <select class="form-select" name="color">
-                    <option value="Black">Black</option>
-                    <option value="White">White</option>
-                    <option value="Green">Green</option>
-                    <option value="Red">Red</option>
-                    <option value="Blue">Blue</option>
-                </select>
-            </div>
-            <div class="mb-3">
-                <label>Kích thước:</label>
-                <select class="form-select" name="size">
-                    <option value="S">S</option>
-                    <option value="M">M</option>
-                    <option value="L">L</option>
-                    <option value="XL">XL</option>
-                    <option value="XXL">XXL</option>
-                </select>
-            </div>
-            <button type="submit" class="btn btn-success">Thêm Sản Phẩm</button>
+            <button type="submit" class="btn btn-success">chỉnh sửa</button>
         </form>
     </div>
-    <script>
-        document.addEventListener("DOMContentLoaded", function() {
-            const sidebar = document.getElementById('sidebar');
-            const content = document.getElementById('content');
-
-            document.getElementById('toggle-btn').addEventListener('click', function() {
-                sidebar.classList.toggle('collapsed');
-                content.classList.toggle('full-width');
-            });
-        });
-    </script>
 </body>
 
 </html>
